requirements.txt
Flask==2.3.3
flask-cors==3.1.2
tensorflow==2.13.0
numpy==1.26.0
Pillow==10.0.0
opencv-python==4.8.1.78
rembg==2.1.0
